+++ src/feature/updateCommentText/index.ts
@@ -1,0 +1,1 @@
+export * from './ui';
